fix: 학습 프래그먼트 예외처리 - 어댑터 연결을 보장하기 위해 먼저 emptyList()로 연결한 후 API가 호출되면 notifySetDataChanged() 사용해서 업데이트 - 무한 스크롤에서 데이터의 개수가 3개 이하일 때 adjustSize와 margin설정에 오류가 생기는 것을 확인하고 예외처리를 강화함

+++ app/src/main/java/kr/dori/android/own_cast/study/StudyAdapter.kt
@@ -49,7 +49,8 @@ class StudyAdapter() : RecyclerView.Adapter<StudyAdapter.Holder>() {
             }
         }
         fun setText(data: PlaylistInfo){
-            Glide.with(itemView.context).load(data.imagePath.toInt()).into(binding.studyCategoryIv)
+            // 이미지 URL을 Glide로 로드
+            Glide.with(itemView.context).load(data.imagePath).into(binding.studyCategoryIv)
             binding.studyCategoryTv.text = data.playlistName
         }
     }
